change runner to accept cb injection for testing

+++ runner.js
@@ -8,59 +8,56 @@ let poller;
 
 let threadCount = 0;
 
-const runTest = test => {
-  try {
-    threadCount++;
-    const startTime = Date.now();
-    const first = spawn(
-      `git clone ${test['PR Url']}.git ${test['Student ID']}${startTime}`,
+const runTest = (test, cb) => {
+  threadCount++;
+  const startTime = Date.now();
+  const first = spawn(
+    `git clone ${test['PR Url']}.git ${test['Student ID']}${startTime}`,
+    {
+      shell: true
+    }
+  );
+  first.on('close', () => {
+    const second = spawn(
+      `cd ${test['Student ID']}${startTime} && yarn install`,
       {
         shell: true
       }
     );
-    first.on('close', () => {
-      const second = spawn(
-        `cd ${test['Student ID']}${startTime} && yarn install`,
+    if (cb) return cb({ ...test, first }); // needed for git test to exit
+    second.on('close', () => {
+      const third = spawn(
+        `cd ${test['Student ID']}${startTime} && yarn test:sis`,
         {
           shell: true
         }
       );
-      second.on('close', () => {
-        const third = spawn(
-          `cd ${test['Student ID']}${startTime} && yarn test:sis`,
+      third.on('close', () => {
+        const testResults = require(`${cwd}/${
+          test['Student ID']
+        }${startTime}/testRun`);
+        const fourth = spawn(
+          `rm -rf ${cwd}/${test['Student ID']}${startTime}`,
           {
             shell: true
           }
         );
-        third.on('close', () => {
-          const testResults = require(`${cwd}/${
-            test['Student ID']
-          }${startTime}/testRun`);
-          const fourth = spawn(
-            `rm -rf ${cwd}/${test['Student ID']}${startTime}`,
-            {
-              shell: true
-            }
-          );
-          fourth.on('close', () => {
-            const runTime = Date.now() - startTime;
-            test['Results'] = JSON.stringify(testResults);
-            test['Pass'] = testResults.numFailedTests === 0;
-            test['Student ID'] = [`${test['Student ID']}`];
-            test['Run Time'] = runTime;
-            threadCount--;
-            console.log(`Test Finished in: ${runTime}ms`);
-            store.setRunTimeLog(runTime);
-            if (!store.inTuneMode()) {
-              newTestResult(test);
-            }
-          });
+        fourth.on('close', () => {
+          const runTime = Date.now() - startTime;
+          test['Results'] = JSON.stringify(testResults);
+          test['Pass'] = testResults.numFailedTests === 0;
+          test['Student ID'] = [`${test['Student ID']}`];
+          test['Run Time'] = runTime;
+          threadCount--;
+          console.log(`Test Finished in: ${runTime}ms`);
+          store.setRunTimeLog(runTime);
+          if (!store.inTuneMode()) {
+            newTestResult(test);
+          }
         });
       });
     });
-  } catch (error) {
-    console.log(error);
-  }
+  });
 };
 
 const runner = () => {
